apply POYO changes to POYOPlus
