+++ examples/poyo/configs/dataset/perich_miller_population_2018.yaml
@@ -127,9 +127,9 @@
       normalize_mean: 0.
       normalize_std: 20.
       weights:
-      movement_phases.random_period: 1.0
-      movement_phases.hold_period: 0.1
-      cursor_outlier_segments: 0.0
+        movement_phases.random_period: 1.0
+        movement_phases.hold_period: 0.1
+        cursor_outlier_segments: 0.0
       metrics:
         - metric:
             _target_: torchmetrics.R2Score
